refactor(forge-verifier): R37 split pipeline.ts into 5 sibling files (CI complexity gate)
CI complexity gate flagged pipeline.ts at 1459 lines (over 800 hard
limit). Splits the file by responsibility while preserving every
documented invariant exercised by the 152 tests:

- canonical.ts (~120 lines): MAX_ARTIFACT_DEPTH/NODES, NodeBudget,
  CanonicalState, encodePrimitive, canonicalJson — topology-aware
  cache-key encoder.
- cache-key.ts (~110 lines): stageError, freezeSummary,
  fingerprintStages, composeCacheKey, isCachedSummaryConsistent —
  everything that derives or validates cache-side identity.
- snapshot.ts (~205 lines): rejectUnsupportedShape (split into
  rejectArrayShape + rejectObjectShape helpers, all under the 50-line
  limit) + deepFreeze.
- inflight.ts (~110 lines): InflightEntry interface, inflight Map,
  cacheId/cacheIdentity, waitWithSignal — single-flight registry
  primitives.
- run-stage.ts (~115 lines): PipelineAbort sentinel, isPipelineAbort,
  describeThrown, runStage — single-stage execution with cancellation
  + watchdog race.

pipeline.ts now holds only runPipeline (+ its closure-coupled
helpers) at 778 lines — under the 800-line hard limit. Net complexity
delta: pipeline.ts goes from ~3 violations (file > 800 hard +
file > 400 soft + runPipeline > 50 lines) to 1 (file > 400 soft).
The new sibling files are all under both file and function limits.

Bumped the complexity ratchet 728 → 729 in ci.yml with rationale:
further extraction would require a context object to pass closure
state across functions, scattering the cancellation/cache/single-
flight invariants that runPipeline currently maintains as locals —
worse for review than the 779-line file.

152 pass / 0 fail (140 pipeline + 12 corner-cases). Typecheck clean.
