+++ packages/lib/forge-verifier/src/pipeline.ts
@@ -1,500 +1,17 @@
-import { types } from "node:util";
-import type {
-  ForgeStageDigest,
-  ForgeVerificationSummary,
-  KoiError,
-  KoiErrorCode,
-  Result,
-} from "@koi/core";
+import type { ForgeStageDigest, ForgeVerificationSummary, KoiErrorCode, Result } from "@koi/core";
 import { RETRYABLE_DEFAULTS } from "@koi/core";
-import type { StageContext, StageOutcome, VerifierStage, VerifyOptions } from "./types.js";
-
-/**
- * Process-local single-flight registry. Two `runPipeline` callers that
- * compute the same `composedKey` concurrently both miss in cache.get,
- * both run the full stage pipeline, then race to write the same key.
- * For pure stages this is wasted CPU; for the explicitly-supported
- * non-idempotent stages (sandbox jobs, external APIs, quota-bearing
- * checks), it is duplicated irreversible work.
- *
- * Coalescing here means the FIRST caller's pipeline runs, and any
- * concurrent caller with the same key awaits its result. Followers
- * inherit the leader's success/failure verbatim — including any
- * leader-side abort. A follower's own AbortSignal cannot interrupt
- * the leader (cooperative cancellation is per-pipeline, not per-key).
- *
- * Process-local only: cross-process deduplication requires backend
- * cooperation (e.g. a Redis SET NX), out of scope here.
- */
-/**
- * `InflightEntry` carries the shared leader promise AND a `detach` callback
- * the orchestrator invokes the moment a SECOND caller joins. The leader's
- * pipeline starts wired to the leader's own `AbortSignal` (so a solo
- * cache-backed run with no follower still aborts on caller cancellation).
- * As soon as a follower attaches, `detach()` unwires the leader's caller
- * signal — from that point the shared work runs to completion regardless
- * of the leader caller's abort, because some other caller is now relying
- * on it. The leader caller continues to see its own TIMEOUT via the outer
- * `waitWithSignal` race; it simply no longer cancels the underlying work.
- */
-interface InflightEntry {
-  readonly promise: Promise<Result<ForgeVerificationSummary>>;
-  readonly detach: () => void;
-  /**
-   * The leader pipeline's internal signal (the one the stage loop
-   * honors). Followers consult it before attaching — if it is already
-   * aborted the leader pipeline is doomed, so a follower that joined
-   * here would receive TIMEOUT despite never having aborted. Late
-   * arrivals run a fresh pipeline instead.
-   */
-  readonly leaderPipelineSignal: AbortSignal | undefined;
-  /**
-   * Registers a follower as a live consumer of the shared result.
-   * Increments the leader's `liveConsumers` counter and (when
-   * `followerSignal` is provided) wires an abort listener that
-   * decrements when the follower itself aborts. Used by the
-   * cache-write gate so a result no live caller actually accepted is
-   * not persisted.
-   */
-  readonly registerConsumer: (followerSignal: AbortSignal | undefined) => void;
-}
-const inflight = new Map<string, InflightEntry>();
-
-/**
- * Per-cache-instance identity for the single-flight key. Two callers
- * coalescing on the same composedKey but DIFFERENT cache backends or
- * DIFFERENT cacheReadFailure policies must not share a leader: a caller
- * with a trusted cache could otherwise inherit a forged pass from a
- * caller with an untrusted cache, and a `"fail"` caller could be dragged
- * into a `"miss"` leader's silent re-execution after a backend outage.
- *
- * Identity uses a `WeakMap` so cache instances that go out of scope do
- * not retain memory; it is per-process and never crosses workers.
- */
-const cacheIdentity = new WeakMap<object, string>();
-let cacheIdSeq = 0;
-function cacheId(cache: object): string {
-  const existing = cacheIdentity.get(cache);
-  if (existing !== undefined) return existing;
-  const id = `c${++cacheIdSeq}`;
-  cacheIdentity.set(cache, id);
-  return id;
-}
-
-/**
- * Race an in-flight leader promise against this caller's own AbortSignal.
- * If the signal fires first, return TIMEOUT to THIS caller — the leader
- * keeps running for other followers. If the leader resolves first, return
- * its result verbatim. Either way, no cross-caller leakage.
- */
-async function waitWithSignal(
-  leader: Promise<Result<ForgeVerificationSummary>>,
-  signal: AbortSignal,
-): Promise<Result<ForgeVerificationSummary>> {
-  if (signal.aborted) {
-    return {
-      ok: false,
-      error: {
-        code: "TIMEOUT",
-        message: "Caller aborted while awaiting in-flight verification.",
-        retryable: RETRYABLE_DEFAULTS.TIMEOUT,
-        context: { stage: "<inflight>" },
-      },
-    };
-  }
-  return new Promise((resolve) => {
-    const onAbort = (): void => {
-      resolve({
-        ok: false,
-        error: {
-          code: "TIMEOUT",
-          message: "Caller aborted while awaiting in-flight verification.",
-          retryable: RETRYABLE_DEFAULTS.TIMEOUT,
-          context: { stage: "<inflight>" },
-        },
-      });
-    };
-    signal.addEventListener("abort", onAbort, { once: true });
-    leader.then(
-      (value) => {
-        signal.removeEventListener("abort", onAbort);
-        resolve(value);
-      },
-      (err: unknown) => {
-        signal.removeEventListener("abort", onAbort);
-        // Leader threw (should be unreachable — work() always returns Result),
-        // but defensively map to INTERNAL inside the envelope rather than
-        // letting the rejection propagate.
-        const detail = err instanceof Error ? err.message : "in-flight verification threw";
-        resolve({
-          ok: false,
-          error: {
-            code: "INTERNAL",
-            message: `Leader pipeline threw: ${detail}`,
-            retryable: RETRYABLE_DEFAULTS.INTERNAL,
-            context: { stage: "<inflight>" },
-            cause: err,
-          },
-        });
-      },
-    );
-  });
-}
-
-function stageError(code: KoiErrorCode, stage: string, message: string, cause?: unknown): KoiError {
-  return {
-    code,
-    message,
-    retryable: RETRYABLE_DEFAULTS[code],
-    context: { stage },
-    ...(cause !== undefined ? { cause } : {}),
-  };
-}
-
-/**
- * Deep-freeze a summary so neither callers nor any cache backend can hand
- * back mutable state. Applied to fresh results before returning AND to every
- * cache hit before it's trusted.
- */
-function freezeSummary(summary: ForgeVerificationSummary): ForgeVerificationSummary {
-  return Object.freeze({
-    passed: summary.passed,
-    sandbox: summary.sandbox,
-    totalDurationMs: summary.totalDurationMs,
-    stageResults: Object.freeze(summary.stageResults.map((d) => Object.freeze({ ...d }))),
-  });
-}
-
-function fingerprintStages<I>(stages: readonly VerifierStage<I>[]): string {
-  // JSON-encode so reserved characters in `name` or `version` cannot collide.
-  // `sandboxed` is part of the identity so flipping a stage from non-sandbox
-  // to sandbox (without bumping `version`) cannot reuse old non-sandbox
-  // cache entries and have them returned as `sandbox: true`.
-  return JSON.stringify(stages.map((s) => [s.name, s.version ?? "0", s.sandboxed === true]));
-}
-
-function composeCacheKey<I>(
-  namespace: string,
-  artifactDigest: string,
-  stages: readonly VerifierStage<I>[],
-  executionContextKey: string | undefined,
-  stageTimeoutMs: number | undefined,
-): string {
-  // JSON-encode each component so reserved characters in any single
-  // component cannot alias a different tuple. The execution-context
-  // key is also folded in so two callers with the same artifact +
-  // stages but different ambient context (auth, tenant policy, etc.)
-  // do not share a cached pass. `stageTimeoutMs` is folded in too:
-  // a permissive caller's success (stage took 500ms under a 1000ms
-  // budget) must NOT satisfy a stricter caller (50ms budget) on a
-  // cache hit — the stage would have timed out under the stricter
-  // policy. Partitioning the key by normalized timeout prevents
-  // policy drift across tenants/environments that share a backend.
-  return JSON.stringify([
-    namespace,
-    artifactDigest,
-    fingerprintStages(stages),
-    executionContextKey ?? "",
-    stageTimeoutMs !== undefined ? String(stageTimeoutMs) : "",
-  ]);
-}
-
-/**
- * Validate that a cached summary actually corresponds to the current stage
- * list AND trust posture. A corrupted, malicious, or stale cache backend
- * can otherwise return an empty `stageResults` array (or one with the wrong
- * stage names) and the pipeline would skip every check while reporting
- * `passed: true`. Also rejects entries whose stored `sandbox` claim differs
- * from the current static declaration — defense in depth: stage `sandboxed`
- * is already part of the cache key, so a divergence here means the backend
- * is replaying across keys it should never satisfy.
- */
-function isFiniteNonNegative(v: unknown): v is number {
-  return typeof v === "number" && Number.isFinite(v) && v >= 0;
-}
-
-function isCachedSummaryConsistent<I>(
-  summary: unknown,
-  stages: readonly VerifierStage<I>[],
-  declaredSandbox: boolean,
-): summary is ForgeVerificationSummary {
-  // Cache backends are explicitly pluggable and may be remote — one
-  // corrupted row, buggy adapter, or cross-version payload must not throw
-  // a TypeError out of `runPipeline` or pass through values that downstream
-  // consumers (e.g. `createForgeProvenance`) reject. Validate every field
-  // including duration shape — NaN/Infinity in a cached row would surface
-  // as a persistence failure later, far from this boundary.
-  if (summary === null || typeof summary !== "object") return false;
-  const s = summary as Record<string, unknown>;
-  if (s.passed !== true) return false;
-  if (s.sandbox !== declaredSandbox) return false;
-  if (!isFiniteNonNegative(s.totalDurationMs)) return false;
-  if (!Array.isArray(s.stageResults)) return false;
-  if (s.stageResults.length !== stages.length) return false;
-  for (let i = 0; i < stages.length; i++) {
-    const expected = stages[i];
-    const got = s.stageResults[i];
-    if (expected === undefined || got === null || typeof got !== "object") return false;
-    const d = got as Record<string, unknown>;
-    if (d.stage !== expected.name) return false;
-    if (d.passed !== true) return false;
-    if (!isFiniteNonNegative(d.durationMs)) return false;
-  }
-  return true;
-}
-
-/**
- * Deterministic JSON of a frozen, validated snapshot. Sorted keys at every
- * level so two structurally-equal artifacts always produce the same string.
- * Throws on cycles — used as the artifact-side cache key, callers without a
- * cycle-tolerant artifact get cache bypass (a correctness-preserving miss),
- * not a false hit. NEVER invokes caller code: by the time we reach here,
- * the snapshot is plain data that was already cloned and frozen.
- */
-/**
- * Encode a primitive leaf for the canonical key. JSON.stringify is not
- * injective for JS values:
- *   - NaN / ±Infinity all serialize to "null"
- *   - -0 serializes to "0"
- *   - undefined yields literal undefined (string-concat hazard)
- *   - bigint throws
- *
- * Bare-string sentinels like `"#NaN"` would collide with user strings of
- * the same content. Every leaf instead gets a TYPE TAG prefix unique to
- * its JS type — `s:`, `f:`, `b:`, `n:`, `u:`, `g:` — so the encoded
- * string for a value of one type can never equal the encoded string for
- * a value of any other type. Arrays/objects retain their `[`/`{` prefix
- * and never start with a tag, so they are distinguishable too.
- */
-function encodePrimitive(value: unknown): string {
-  if (value === null) return "n:";
-  if (value === undefined) return "u:";
-  if (typeof value === "boolean") return value ? "b:t" : "b:f";
-  if (typeof value === "number") {
-    if (Number.isNaN(value)) return "f:NaN";
-    if (value === Number.POSITIVE_INFINITY) return "f:+Inf";
-    if (value === Number.NEGATIVE_INFINITY) return "f:-Inf";
-    if (Object.is(value, -0)) return "f:-0";
-    return `f:${value}`;
-  }
-  if (typeof value === "string") return `s:${JSON.stringify(value)}`;
-  if (typeof value === "bigint") return `g:${value.toString()}`;
-  // Symbols/functions cannot reach here — rejectUnsupportedShape rejects
-  // them upstream — but keep a defensive tag rather than fall through.
-  return `?:${JSON.stringify(String(value))}`;
-}
-
-/**
- * Hard caps on artifact graph traversal. Both `rejectUnsupportedShape`
- * and `canonicalJson` recurse synchronously over attacker-controlled
- * input; an abort signal cannot interrupt synchronous JS, so the only
- * way to bound preprocessing CPU under cancellation pressure is to bound
- * the input itself.
- *
- *   - DEPTH (256): catches stack-blowing inputs; safely below V8's default
- *     call-stack limit while exceeding any realistic config artifact.
- *   - NODES (50_000): catches wide flat inputs (e.g. an array/object with
- *     1M entries) that the depth cap alone cannot stop. Counts every
- *     visited primitive AND every container so the bound is total work,
- *     not just leaf work.
- */
-const MAX_ARTIFACT_DEPTH = 256;
-const MAX_ARTIFACT_NODES = 50_000;
-
-interface NodeBudget {
-  count: number;
-}
-
-interface CanonicalState {
-  readonly onStack: WeakSet<object>;
-  readonly seen: WeakMap<object, number>;
-  readonly budget: NodeBudget;
-  refCounter: number;
-}
-
-/**
- * Topology-aware canonical serializer. Tracks two things:
- *
- *   - `onStack`: ancestors of the current node — a re-entry is a TRUE
- *     cycle and throws (no deterministic linearization exists).
- *   - `seen`: every previously-visited object → integer ID. Re-encountering
- *     a shared subobject (DAG aliasing) emits `ref:N` instead of recursing.
- *     This makes shared-reference DAGs cacheable AND distinct from
- *     identical-content non-shared graphs: stages observe reference
- *     identity (`a.x === a.y`), so the cached pass is bound to topology.
- */
-function canonicalJson(value: unknown, state: CanonicalState, depth = 0): string {
-  if (depth > MAX_ARTIFACT_DEPTH) {
-    throw new Error(`snapshot exceeds maximum depth (${MAX_ARTIFACT_DEPTH})`);
-  }
-  state.budget.count += 1;
-  if (state.budget.count > MAX_ARTIFACT_NODES) {
-    throw new Error(`snapshot exceeds maximum node count (${MAX_ARTIFACT_NODES})`);
-  }
-  if (value === null || typeof value !== "object") return encodePrimitive(value);
-  if (state.onStack.has(value)) {
-    // Tagged so callers can distinguish "expected: cyclic snapshot" from
-    // serializer bugs / future regressions and decide bypass-vs-fail.
-    const err = new Error("snapshot contains a cycle; cannot derive a deterministic cache key");
-    (err as Error & { code?: string }).code = "FORGE_VERIFIER_CYCLE";
-    throw err;
-  }
-  // Topology-aware aliasing: a node reached via a non-back-edge that we've
-  // already serialized once gets a stable reference ID. Different DAG
-  // topologies produce different keys; identical-content non-shared graphs
-  // never collide with shared-reference graphs.
-  const priorId = state.seen.get(value);
-  if (priorId !== undefined) {
-    return `ref:${priorId}`;
-  }
-  const id = state.refCounter++;
-  state.seen.set(value, id);
-  state.onStack.add(value);
-  try {
-    if (Array.isArray(value)) {
-      // Iterate 0..length-1 explicitly so a sparse array (`new Array(1)`)
-      // does NOT serialize equal to `[]`. `Array.prototype.map` skips holes,
-      // which would alias different artifact shapes to the same cache key.
-      const parts: string[] = [];
-      for (let i = 0; i < value.length; i++) {
-        parts.push(canonicalJson(value[i], state, depth + 1));
-      }
-      return `#${id}[${parts.join(",")}]`;
-    }
-    const obj = value as Record<string, unknown>;
-    const keys = Object.keys(obj).sort();
-    const parts = keys.map(
-      (k) => `${JSON.stringify(k)}:${canonicalJson(obj[k], state, depth + 1)}`,
-    );
-    return `#${id}{${parts.join(",")}}`;
-  } finally {
-    state.onStack.delete(value);
-  }
-}
-
-/**
- * Sentinel raised when the caller signal aborts mid-stage OR a per-stage
- * watchdog elapses. Distinguished from real stage throws so the caller
- * gets a TIMEOUT outcome rather than INTERNAL.
- */
-const ABORT_BY_PIPELINE = Symbol("forge-verifier:abort");
-class PipelineAbort extends Error {
-  readonly [ABORT_BY_PIPELINE] = true;
-  constructor(reason: string) {
-    super(reason);
-  }
-}
-function isPipelineAbort(e: unknown): e is PipelineAbort {
-  return (
-    e instanceof Error && (e as Error & { [ABORT_BY_PIPELINE]?: true })[ABORT_BY_PIPELINE] === true
-  );
-}
-
-async function runStage<I>(
-  stage: VerifierStage<I>,
-  artifact: I,
-  ctx: StageContext,
-  signal: AbortSignal | undefined,
-  stageTimeoutMs: number | undefined,
-): Promise<{
-  readonly outcome: StageOutcome;
-  readonly durationMs: number;
-  readonly thrown?: unknown;
-  readonly aborted?: true;
-  /**
-   * The underlying stage promise. The caller's `runPipeline` retains
-   * this so the in-flight slot can be held until the underlying work
-   * settles, even if the caller-visible Promise.race already resolved
-   * with TIMEOUT — otherwise a retry could start a second stage while
-   * the first one is still running, double-submitting irreversible
-   * work for non-idempotent stages.
-   */
-  readonly underlying: Promise<unknown>;
-}> {
-  const started = performance.now();
-  // Capture handles so they can be cleaned up after the race settles —
-  // otherwise every successful stage leaves a live setTimeout AND an
-  // abort listener attached to the caller signal until the timeout
-  // expires or the signal eventually fires. Under load that grows into
-  // a memory leak and `MaxListenersExceededWarning`.
-  let timeoutHandle: ReturnType<typeof setTimeout> | undefined;
-  let abortListener: (() => void) | undefined;
-  // Capture the underlying stage promise BEFORE racing — Promise.race
-  // can't kill it, but we can keep a reference so callers can wait on
-  // its settlement (silently) before releasing the in-flight slot.
-  // Defer the `stage.run` invocation through a microtask so a SYNC
-  // throw from a buggy plugin is normalized into a rejected promise
-  // rather than escaping `runPipeline` and breaking the documented
-  // `Promise<Result<...>>` contract.
-  // Re-check the signal INSIDE the microtask too — a caller can abort
-  // in the gap between the outer loop's pre-stage gate and this
-  // microtask firing; without this re-check, the stage would still
-  // start and could produce a side effect after the caller already
-  // gave up.
-  const underlying = Promise.resolve().then(() => {
-    if (signal?.aborted === true) {
-      throw new PipelineAbort("aborted before stage start (microtask race)");
-    }
-    return stage.run(artifact, ctx);
-  });
-  try {
-    // Race stage execution against the caller signal AND the optional
-    // per-stage watchdog. A buggy or hostile plugin that ignores
-    // ctx.signal cannot wedge the pipeline beyond stageTimeoutMs (or
-    // beyond the caller's signal). The underlying Promise may keep
-    // running — JS gives no way to kill it — but the caller is unblocked.
-    const racers: Promise<StageOutcome>[] = [underlying];
-    if (signal !== undefined) {
-      racers.push(
-        new Promise<StageOutcome>((_, reject) => {
-          if (signal.aborted) {
-            reject(new PipelineAbort("aborted via signal"));
-            return;
-          }
-          abortListener = (): void => reject(new PipelineAbort("aborted via signal"));
-          signal.addEventListener("abort", abortListener, { once: true });
-        }),
-      );
-    }
-    if (stageTimeoutMs !== undefined && Number.isFinite(stageTimeoutMs) && stageTimeoutMs > 0) {
-      racers.push(
-        new Promise<StageOutcome>((_, reject) => {
-          timeoutHandle = setTimeout(
-            () =>
-              reject(
-                new PipelineAbort(
-                  `stage exceeded stageTimeoutMs=${stageTimeoutMs}ms (uncooperative plugin?)`,
-                ),
-              ),
-            stageTimeoutMs,
-          );
-        }),
-      );
-    }
-    const outcome = await Promise.race(racers);
-    return { outcome, durationMs: performance.now() - started, underlying };
-  } catch (e: unknown) {
-    if (isPipelineAbort(e)) {
-      return {
-        outcome: { ok: false, reason: e.message },
-        durationMs: performance.now() - started,
-        aborted: true,
-        underlying,
-      };
-    }
-    return {
-      outcome: { ok: false, reason: "stage threw", cause: e },
-      durationMs: performance.now() - started,
-      thrown: e,
-      underlying,
-    };
-  } finally {
-    if (timeoutHandle !== undefined) clearTimeout(timeoutHandle);
-    if (abortListener !== undefined && signal !== undefined) {
-      signal.removeEventListener("abort", abortListener);
-    }
-  }
-}
+import {
+  composeCacheKey,
+  fingerprintStages,
+  freezeSummary,
+  isCachedSummaryConsistent,
+  stageError,
+} from "./cache-key.js";
+import { canonicalJson } from "./canonical.js";
+import { cacheId, inflight, waitWithSignal } from "./inflight.js";
+import { describeThrown, runStage } from "./run-stage.js";
+import { deepFreeze, rejectUnsupportedShape } from "./snapshot.js";
+import type { StageContext, VerifierStage, VerifyOptions } from "./types.js";
 
 /**
  * Sequential verification orchestrator.
@@ -1257,202 +774,4 @@ export async function runPipeline<I>(
   // Solo path: still clean up listeners after work() settles so the
   // leader signal is not retained beyond the verification.
   return work().finally(releaseConsumerListeners);
-}
-
-/**
- * Reject artifacts that contain types whose mutability cannot be enforced by
- * `Object.freeze`, or whose snapshot would silently differ from the original.
- *
- * - Typed arrays / DataView / ArrayBuffer: `Object.freeze` throws on populated
- *   typed arrays in V8, and a frozen typed array's underlying buffer is still
- *   mutable. We do not support binary blobs in the artifact graph yet.
- * - `Map` / `Set`: `Object.freeze` does NOT block `.set` / `.add` / `.delete`,
- *   so a frozen collection is still mutable at the API surface. Reject
- *   instead of pretending it's immutable.
- * - Class instances (non-plain objects): `structuredClone` silently strips the
- *   prototype, so stages would receive a plain-object that is not equal to
- *   what the caller passed in and the cached pass would attest to a different
- *   shape than the input.
- */
-function rejectUnsupportedShape(
-  value: unknown,
-  path: string,
-  seen: WeakSet<object>,
-  budget: NodeBudget,
-  depth = 0,
-): void {
-  if (depth > MAX_ARTIFACT_DEPTH) {
-    throw new TypeError(
-      `Artifact at ${path} exceeds maximum depth (${MAX_ARTIFACT_DEPTH}); deeply-nested artifacts are rejected to bound preprocessing CPU under cancellation.`,
-    );
-  }
-  budget.count += 1;
-  if (budget.count > MAX_ARTIFACT_NODES) {
-    throw new TypeError(
-      `Artifact at ${path} exceeds maximum node count (${MAX_ARTIFACT_NODES}); wide artifacts are rejected to bound preprocessing CPU under cancellation.`,
-    );
-  }
-  if (value === null || typeof value !== "object") {
-    // Functions and symbols nested inside a graph are also rejected here
-    // (typeof "object" excludes them). structuredClone would throw, but
-    // catching it here gives a much better path-rooted error message.
-    if (typeof value === "function" || typeof value === "symbol") {
-      throw new TypeError(
-        `Artifact at ${path} has unsupported type "${typeof value}"; verifier requires plain-data artifacts.`,
-      );
-    }
-    return;
-  }
-  // Proxy detection BEFORE any reflective op. `types.isProxy` is a privileged
-  // V8 introspection (Bun + Node both expose it via node:util) that does NOT
-  // invoke any handler. A Proxy artifact is rejected here without ever firing
-  // a trap, closing the only remaining caller-code-on-verifier-stack path.
-  if (types.isProxy(value)) {
-    throw new TypeError(
-      `Artifact at ${path} is a Proxy; verifier requires plain-data artifacts (Proxy traps would execute caller code on the verifier stack).`,
-    );
-  }
-  // Cycle guard: a self-referential plain object is legal for
-  // `structuredClone`, but unbounded recursion here would let a crafted
-  // artifact knock the verifier over. Validate each object exactly once.
-  if (seen.has(value)) return;
-  seen.add(value);
-  if (
-    ArrayBuffer.isView(value) ||
-    value instanceof ArrayBuffer ||
-    value instanceof Map ||
-    value instanceof Set
-  ) {
-    throw new TypeError(
-      `Artifact at ${path} contains ${value.constructor.name}; verifier requires plain-data artifacts (Map/Set/typed-array unsupported).`,
-    );
-  }
-  if (Array.isArray(value)) {
-    // Use descriptor walk for ALL own keys, including numeric indices, so a
-    // hostile array with `Object.defineProperty(arr, "0", { get() { ... } })`
-    // is rejected without firing the getter. Also rejects symbol keys,
-    // accessors, non-enumerable, and extra named props.
-    const arrSymbols = Object.getOwnPropertySymbols(value);
-    if (arrSymbols.length > 0) {
-      throw new TypeError(
-        `Artifact at ${path} (array) has symbol-keyed own properties; verifier requires plain-data artifacts.`,
-      );
-    }
-    // Charge the array's declared length toward the budget BEFORE any
-    // hole scan. Without this, `new Array(1_000_000_000)` forces an
-    // O(length) synchronous scan even though the budget should reject
-    // it instantly.
-    budget.count += value.length;
-    if (budget.count > MAX_ARTIFACT_NODES) {
-      throw new TypeError(
-        `Artifact at ${path} declares length=${value.length} which alone exceeds the maximum node count (${MAX_ARTIFACT_NODES}); arrays are bounded by their declared length to prevent CPU exhaustion before any hole scan.`,
-      );
-    }
-    const arrDescs = Object.getOwnPropertyDescriptors(value);
-    // Reject sparse arrays: every index in [0, length) MUST have an own
-    // data descriptor. A hole would be skipped by Array.prototype.map and
-    // alias to a denser array's cache key — different content, same key.
-    // Bounded by the budget charge above so a hostile huge `length` is
-    // already rejected before we get here.
-    for (let i = 0; i < value.length; i++) {
-      if (!Object.hasOwn(arrDescs, String(i))) {
-        throw new TypeError(
-          `Artifact at ${path}[${i}] is a hole; verifier rejects sparse arrays (holes are skipped by serializers and would alias dense arrays in the cache key).`,
-        );
-      }
-    }
-    for (const [k, desc] of Object.entries(arrDescs)) {
-      if (k === "length") continue;
-      if (typeof desc.get === "function" || typeof desc.set === "function") {
-        throw new TypeError(
-          `Artifact at ${path}.${k} (array property) is an accessor (getter/setter); verifier rejects accessors so caller code never executes during validation.`,
-        );
-      }
-      if (desc.enumerable !== true) {
-        throw new TypeError(
-          `Artifact at ${path}.${k} (array property) is non-enumerable; verifier requires plain-data artifacts.`,
-        );
-      }
-      if (!/^\d+$/.test(k)) {
-        // Extra named properties on arrays are not preserved by
-        // structuredClone, so reject rather than silently drop.
-        throw new TypeError(
-          `Artifact at ${path}.${k} (array property) is a non-index own property; verifier requires plain-data arrays (extra named properties are not preserved by structuredClone).`,
-        );
-      }
-      rejectUnsupportedShape(desc.value, `${path}[${k}]`, seen, budget, depth + 1);
-    }
-    return;
-  }
-  // Non-plain object (class instance, etc.): structuredClone strips the
-  // prototype. Reject so the cached pass cannot attest to a different shape
-  // than what the caller actually passed in.
-  const proto = Object.getPrototypeOf(value);
-  if (proto !== Object.prototype && proto !== null) {
-    throw new TypeError(
-      `Artifact at ${path} is a non-plain object (${proto.constructor.name}); verifier requires plain-data artifacts.`,
-    );
-  }
-  // Reject any symbol-keyed own properties — structuredClone drops them, so
-  // they would never appear in the snapshot, the cache key, or what stages
-  // see, even though the caller's artifact carries them.
-  const symbols = Object.getOwnPropertySymbols(value);
-  if (symbols.length > 0) {
-    throw new TypeError(
-      `Artifact at ${path} has symbol-keyed own properties; verifier requires plain-data artifacts (symbol keys are not preserved by structuredClone).`,
-    );
-  }
-  // Preflight key count BEFORE materializing every descriptor — a
-  // wide attacker-controlled object would otherwise force a full
-  // O(n) descriptor allocation prior to the budget check. Use
-  // Object.getOwnPropertyNames so non-enumerable keys are also counted
-  // and inspected (the non-enumerable rejection runs in the loop below).
-  const keys = Object.getOwnPropertyNames(value);
-  budget.count += keys.length;
-  if (budget.count > MAX_ARTIFACT_NODES) {
-    throw new TypeError(
-      `Artifact at ${path} has ${keys.length} own keys which alone exceeds the maximum node count (${MAX_ARTIFACT_NODES}); wide objects are bounded before descriptor materialization to prevent CPU exhaustion.`,
-    );
-  }
-  // Per-key descriptor walk (one at a time, NOT a bulk
-  // getOwnPropertyDescriptors call) so getters are NOT invoked AND a
-  // wide object cannot force bulk allocation. Reject accessors and
-  // non-enumerable own properties outright.
-  for (const k of keys) {
-    const desc = Object.getOwnPropertyDescriptor(value, k);
-    if (desc === undefined) continue;
-    if (typeof desc.get === "function" || typeof desc.set === "function") {
-      throw new TypeError(
-        `Artifact at ${path}.${k} is an accessor (getter/setter); verifier rejects accessors so caller code never executes during validation.`,
-      );
-    }
-    if (desc.enumerable !== true) {
-      throw new TypeError(
-        `Artifact at ${path}.${k} is non-enumerable; verifier requires plain-data artifacts (non-enumerable properties are not preserved by structuredClone).`,
-      );
-    }
-    rejectUnsupportedShape(desc.value, `${path}.${k}`, seen, budget, depth + 1);
-  }
-}
-
-/**
- * Recursively freeze plain objects and arrays. Map/Set/typed arrays are
- * rejected upstream by `rejectUnsupportedShape`, so we never reach them here.
- * Already-frozen substructures are skipped to avoid redundant work.
- */
-function deepFreeze(value: unknown): void {
-  if (value === null || typeof value !== "object") return;
-  if (Object.isFrozen(value)) return;
-  Object.freeze(value);
-  if (Array.isArray(value)) {
-    for (const item of value) deepFreeze(item);
-    return;
-  }
-  for (const v of Object.values(value)) deepFreeze(v);
-}
-
-function describeThrown(thrown: unknown): string {
-  if (thrown instanceof Error) return thrown.message;
-  if (typeof thrown === "string") return thrown;
-  return "non-Error throw";
 }
